reset field handling
